mcux: Update to MCUXpresso SDK 2.7.0
Updates the MCUXpresso SDK to version 2.7.0 (2019-12-19) for all
available SoCs. The KL25Z and KW SoCs are not updated here because they
were not included in the MCUXpresso SDK 2.7.0 release.

Origin: MCUXpresso SDK
License: BSD 3-Clause
URL: mcux.nxp.com
Maintained-by: External

+++ mcux/devices/LPC54114/fsl_power.h
@@ -93,7 +93,7 @@ extern "C" {
 static inline void POWER_EnablePD(pd_bit_t en)
 {
     /* PDRUNCFGSET */
-    SYSCON->PDRUNCFGSET[(en >> 8UL)] = (1UL << (en & 0xffU));
+    SYSCON->PDRUNCFGSET[((uint32_t)en >> 8UL)] = (1UL << ((uint32_t)en & 0xffU));
 }
 
 /*!
@@ -105,7 +105,7 @@ static inline void POWER_EnablePD(pd_bit_t en)
 static inline void POWER_DisablePD(pd_bit_t en)
 {
     /* PDRUNCFGCLR */
-    SYSCON->PDRUNCFGCLR[(en >> 8UL)] = (1UL << (en & 0xffU));
+    SYSCON->PDRUNCFGCLR[((uint32_t)en >> 8UL)] = (1UL << ((uint32_t)en & 0xffU));
 }
 
 /*!
